Small changes to docstrings
